+++ packages/angular/src/lib/components/chat/copilot-chat-input.ts
@@ -156,8 +156,7 @@ export interface ToolbarContext {
           <div class="flex items-center">
             @if (computedMode() === "transcribe") {
               @if (
-                cancelTranscribeButtonTemplate ||
-                cancelTranscribeButtonComponent()
+                cancelTranscribeButtonTemplate || cancelTranscribeButtonComponent()
               ) {
                 <copilot-slot
                   [slot]="
@@ -176,8 +175,7 @@ export interface ToolbarContext {
                 </copilot-chat-cancel-transcribe-button>
               }
               @if (
-                finishTranscribeButtonTemplate ||
-                finishTranscribeButtonComponent()
+                finishTranscribeButtonTemplate || finishTranscribeButtonComponent()
               ) {
                 <copilot-slot
                   [slot]="
@@ -197,8 +195,7 @@ export interface ToolbarContext {
               }
             } @else {
               @if (
-                startTranscribeButtonTemplate ||
-                startTranscribeButtonComponent()
+                startTranscribeButtonTemplate || startTranscribeButtonComponent()
               ) {
                 <copilot-slot
                   [slot]="