+++ packages/voice/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @copilotkit/voice
 
+## 1.55.0-next.8
+
+### Patch Changes
+
+- Updated dependencies [8aafcbe]
+  - @copilotkit/runtime@1.55.0-next.8
+
 ## 1.55.0-next.7
 
 ### Minor Changes